PBF: more conversions for right joystick uint8 to float (#963)

+++ SerialPrograms/Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStoryTools.h
@@ -514,7 +514,7 @@ void move_camera_until_yolo_object_detected(
     ProControllerContext& context, 
     YOLOv5Detector& yolo_detector, 
     const std::string& target_label,
-    uint8_t initial_x_move,
+    double initial_x_move,
     Milliseconds initial_hold,
     uint16_t max_rounds = 50
 );
